FinAgent/agent_backend/requirements.txt
openai-agents==0.0.7
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.104.1